upgraded mkdirs.sh

+++ experimental/w01/rmdirs.sh
@@ -1,3 +1,7 @@
-for i in $(cat $1); do
-	rmdir $i
-done
+if [ -z $1 ];then
+	echo "Enter a txt file to extract words from"
+else
+	for i in $(cat $1); do
+		rmdir $i
+	done
+fi
